app_requirements.txt
--find-links https://pypi.python.org/simple/index.html
--extra-index-url https://pypi.python.org/simple
adal==1.2.7
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
applicationinsights==0.11.10
arch==5.6.0
argcomplete==3.6.3
asttokens==3.0.1
attrs==25.4.0
azure-ai-ml==1.24.0
azure-common==1.1.28
azure-core==1.36.0
azure-graphrbac==0.61.2
azure-identity==1.17.0
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==14.0.0
azure-mgmt-core==1.6.0
azure-mgmt-keyvault==11.0.0
azure-mgmt-network==30.0.0
azure-mgmt-resource==24.0.0
azure-mgmt-storage==24.0.0
azure-monitor-opentelemetry-exporter==1.0.0b40
azure-storage-blob==12.28.0
azure-storage-file-datalake==12.19.0
azure-storage-file-share==12.20.0
azure-storage-queue==12.14.1
azureml-automl-core==1.61.0.post1
azureml-automl-runtime==1.61.0
azureml-core==1.61.0.post1
azureml-dataprep==5.4.2
azureml-dataprep-native==42.1.0
azureml-dataprep-rslex==2.25.2
azureml-dataset-runtime==1.61.0
azureml-defaults==1.61.0
azureml-inference-server-http==1.5.0
azureml-interpret==1.61.0
azureml-mlflow==1.61.0.post1
azureml-pipeline-core==1.61.0
azureml-responsibleai==1.61.0
azureml-telemetry==1.61.0
azureml-train-automl==1.61.0.post1
azureml-train-automl-client==1.61.0.post1
azureml-train-automl-runtime==1.61.0
azureml-train-core==1.61.0
azureml-train-restclients-hyperdrive==1.61.0
azureml-training-tabular==1.61.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==5.0.0
blinker==1.9.0
bokeh==2.4.3
boto3==1.35.0
botocore==1.35.0
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.1
cmdstanpy==1.0.4
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contextlib2==21.6.0
contourpy==1.3.2
convertdate==2.4.0
cryptography==45.0.7
cycler==0.12.1
Cython==3.2.2
dask==2023.2.0
databricks-sdk==0.74.0
dataclasses==0.6
debugpy==1.8.18
decorator==5.2.1
Deprecated==1.3.1
dice-ml==0.11
dill==0.3.9
distributed==2023.2.0
distro==1.9.0
docker==7.1.0
# dotnetcore2==3.1.23  # yanked (license violation) + downloads .NET runtime at install time — causes Docker build to hang
econml==0.16.0
entrypoints==0.4
ephem==4.2
erroranalysis==0.5.5
exceptiongroup==1.3.1
executing==2.2.1
fairlearn==0.7.0
fire==0.7.1
fixedint==0.1.6
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.61.0
fsspec==2023.1.0
fusepy==3.0.1
gensim==4.3.2
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
graphviz==0.21
gunicorn==23.0.0
h11==0.14.0
holidays==0.46
httpcore==1.0.8
httpx==0.27.2
humanfriendly==10.0
idna==3.11
importlib_metadata==7.2.1
importlib_resources==6.4.0
inference-schema==1.8
interpret-core==0.5.0
interpret_community==0.31.0
ipykernel==7.1.0
ipython==8.37.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jiter==0.8.0
jmespath==1.0.1
joblib==1.2.0
jsonpickle==4.1.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
keras2onnx==1.6.0
kiwisolver==1.4.9
knack==0.12.0
kneed==0.8.6
lightgbm==4.6.0
llvmlite==0.39.1
locket==1.0.0
LunarCalendar==0.0.9
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
matplotlib-inline==0.2.1
mealy==0.2.5
ml_wrappers==0.5.6
mlflow-skinny==2.15.1
mltable==1.6.3
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
msgpack==1.1.2
msrest==0.7.1
msrestazure==0.6.4.post1
narwhals==2.18.1
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
networkx==2.5
numba==0.56.4
numpy==1.23.5
oauthlib==3.3.1
onnx==1.16.1
onnxconverter-common==1.13.0
onnxmltools==1.12.0
onnxruntime==1.16.3
openai==1.77.0
opentelemetry-api==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
packaging==24.2
pandas==1.5.3
paramiko==3.5.1
parso==0.8.5
partd==1.4.2
pathspec==0.12.1
patsy==1.0.2
pillow==12.0.0
pkginfo==1.12.1.2
platformdirs==4.5.1
pmdarima==1.8.5
prompt_toolkit==3.0.52
property-cached==1.6.4
prophet==1.1.4
protobuf==5.29.5
psutil==5.9.8
pure_eval==0.2.3
pyarrow==14.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.12.0
pydantic_core==2.33.2
pydash==6.0.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
PyMeeus==0.5.12
PyNaCl==1.6.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
pyreadline3==3.4.1; sys_platform == "win32"  # Windows-only — skipped automatically on Linux/Docker
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2024.2
pywin32==306; sys_platform == "win32"  # Windows-only — skipped automatically on Linux/Docker
PyYAML==6.0.3
pyzmq==27.1.0
raiutils==0.4.2
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
responsibleai==0.36.0
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml==0.18.6
s3transfer==0.10.4
scikit-learn==1.5.1
scipy==1.11.0
seaborn==0.13.2
semver==2.13.0
shap==0.49.1
six==1.17.0
skl2onnx==1.15.0
sklearn-pandas==1.7.0
slicer==0.0.7
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sparse==0.17.0
sqlparse==0.5.4
stack-data==0.6.3
starlette==0.50.0
statsmodels==0.13.5
streamlit==1.56.0
strictyaml==1.7.3
sympy==1.13.1
tabulate==0.9.0
tblib==3.2.2
tenacity==9.1.4
termcolor==3.2.0
threadpoolctl==3.6.0
toml==0.10.2
toolz==1.1.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.12.2
ujson==5.11.0
urllib3==2.5.0
waitress==3.0.2
watchdog==6.0.0
wcwidth==0.2.14
Werkzeug==3.1.4
wrapt==1.16.0
xgboost==1.5.2
zict==3.0.0
zipp==3.23.0
# LangChain / LangGraph (agent rewrite)
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.15
langchain-core==1.2.26
langchain-openai==1.1.12
langsmith==0.7.25
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.12
requests-toolbelt==1.0.0
tiktoken==0.12.0
uuid-utils==0.14.1
ormsgpack==1.12.2
orjson==3.10.18
xxhash==3.5.0

